[update] 크리티컬 계산을 위해 float로 파일에서 입력을 받는것으로 수정 배틀레이팅 계산 추가 데이터 수정 누락된 메타데이터, 아이템 데이터 수정분 추가

+++ Assets/IdleHeaven/Scripts/Effect/ItemEffectData.cs
@@ -65,15 +65,15 @@ namespace IdleHeaven
         [SerializeField] private int _effectId;
         [SerializeField] private StatType _stat;
         [SerializeField] private Rarity _rarity;
-        [SerializeField] private int _minValue;
-        [SerializeField] private int _maxValue;
+        [SerializeField] private float _minValue;
+        [SerializeField] private float _maxValue;
         [SerializeField] private float _levelMultiplier;
         [SerializeField] private float _maxLevelMultiplier;
         public int EffectId { get { return _effectId; } set { _effectId = value; } }
         public StatType Stat { get { return _stat; } set { _stat = value; } }
         public Rarity Rarity { get { return _rarity; } set { _rarity = value; } }
-        public int MinValue { get { return _minValue; } set { _minValue = value; } }
-        public int MaxValue { get { return _maxValue; } set { _maxValue = value; } }
+        public float MinValue { get { return _minValue; } set { _minValue = value; } }
+        public float MaxValue { get { return _maxValue; } set { _maxValue = value; } }
         public float LevelMultiplier { get { return _levelMultiplier; } set { _levelMultiplier = value; } }
         public float MaxLevelMultiplier { get { return _maxLevelMultiplier; } set { _maxLevelMultiplier = value; } }
 
@@ -87,10 +87,10 @@ namespace IdleHeaven
     {
         [SerializeField] private StatType _stat;
         [SerializeField] private Rarity _rarity;
-        [SerializeField] private int _value;
+        [SerializeField] private float _value;
         public StatType Stat { get { return _stat; } set { _stat = value; } }
         public Rarity Rarity { get { return _rarity; } set { _rarity = value; } }
-        public int Value { get { return _value; } set { _value = value; } }
+        public float Value { get { return _value; } set { _value = value; } }
 
         public Color GetRarityColor()
         {
